ARCH-1: extract find dispatch arm into commands::find::run (iter-225)
